attempt to split code into packages
