+++ darwin/darwin_test.go
@@ -1,4 +1,4 @@
-package main
+package darwin
 
 import (
 	"testing"
