--- linux.go
@@ -1,7 +1,0 @@
-package main
-
-func GetLinuxReport() HwReport {
-	return HwReport{}
-}
-
-
