+++ linux/linux.go
@@ -1,0 +1,9 @@
+package linux
+
+import . "github.com/1okey/hw-report/types"
+
+func GetLinuxReport() HwReport {
+	return HwReport{}
+}
+
+
