+++ main.go
@@ -4,6 +4,11 @@ import (
 	"errors"
 	"fmt"
 	"runtime"
+
+	. "github.com/1okey/hw-report/types"
+	. "github.com/1okey/hw-report/darwin"
+	. "github.com/1okey/hw-report/linux"
+	. "github.com/1okey/hw-report/windows"
 )
 
 func get_report() (HwReport, error) {
