+++ types/types.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 type Vendor struct {
 	Name  string
