--- windows.go
@@ -1,5 +1,0 @@
-package main
-
-func GetWinReport() HwReport {
-	return HwReport{}
-}
