+++ windows/windows.go
@@ -1,0 +1,7 @@
+package windows
+
+import . "github.com/1okey/hw-report/types"
+
+func GetWinReport() HwReport {
+	return HwReport{}
+}
